requirements.txt
# Core dependencies for Advanced Rubik's Cube Simulator
# Install with: pip install -r requirements.txt

# Graphics and rendering
numpy>=1.21.0,<2.0.0
pygame>=2.1.0,<3.0.0
PyOpenGL>=3.1.0,<4.0.0
moderngl>=5.6.0,<6.0.0
pillow>=9.0.0,<11.0.0

# Math and algorithms
scipy>=1.8.0,<2.0.0
numba>=0.56.0,<1.0.0

# Configuration and data handling
pydantic>=1.10.0,<3.0.0
pyyaml>=6.0,<7.0.0
toml>=0.10.0,<1.0.0

# CLI and utilities
click>=8.0.0,<9.0.0
rich>=12.0.0,<14.0.0
loguru>=0.6.0,<1.0.0

# Web server and API
fastapi>=0.104.0,<1.0.0
uvicorn>=0.24.0,<1.0.0
websockets>=12.0,<13.0.0
jinja2>=3.1.0,<4.0.0